修改jar包应用

+++ open-home-app/open-home-service-impl/src/main/java/com/shihui/openpf/home/service/impl/OrderServiceImpl.java
@@ -1,19 +1,19 @@
 package com.shihui.openpf.home.service.impl;
 
-import com.shihui.api.common.model.OrderStatusEnum;
-import com.shihui.openpf.home.dao.OrderDao;
-import com.shihui.openpf.home.dao.OrderHistoryDao;
-import com.shihui.openpf.home.model.OrderHistory;
-
-
-import com.shihui.openpf.home.model.Order;
-import com.shihui.openpf.home.model.OrderCancelType;
-import com.shihui.openpf.home.service.api.OrderService;
-import org.springframework.stereotype.Service;
-
-import javax.annotation.Resource;
 import java.util.Date;
 import java.util.List;
+
+import javax.annotation.Resource;
+
+import org.springframework.stereotype.Service;
+
+import com.shihui.api.order.common.enums.OrderStatusEnum;
+import com.shihui.openpf.home.dao.OrderDao;
+import com.shihui.openpf.home.dao.OrderHistoryDao;
+import com.shihui.openpf.home.model.Order;
+import com.shihui.openpf.home.model.OrderCancelType;
+import com.shihui.openpf.home.model.OrderHistory;
+import com.shihui.openpf.home.service.api.OrderService;
 
 /**
  * Created by zhoutc on 2016/1/25.
@@ -68,7 +68,7 @@ public class OrderServiceImpl implements OrderService {
     public boolean updateOrder(long orderId, OrderStatusEnum orderStatus) {
         Order order = new Order();
         order.setOrderId(orderId);
-        order.setOrderStatus((byte)orderStatus.getValue());
+        order.setOrderStatus(orderStatus.getValue());
         if(orderDao.update(order)>0) {
             Date date = new Date();
             OrderHistory orderHistory = new OrderHistory();
